fix: bug where password was rehashed unintentionally and the account was no logner accessible

+++ controllers/profilePicController.js
@@ -6,10 +6,16 @@ const addProfilePic = async (req, res) => {
     const username = req.user.username
 
     try {
-        const image_url = await uploadFileToGoogleCS(username, req.file, 'profPics')
-        
+
         const user = await User.findOne({ username })
-        console.log(user)
+
+        // delete previous profilePic if it was not the default
+        if ( user.profilePic !== 'https://storage.googleapis.com/spatakioutis_app_img/profilePics/default_pic.png') {
+            await deleteFileFromGoogleCS(filepath[filepath.length - 1], 'profPics')
+        }
+        
+        const image_url = await uploadFileToGoogleCS(username, req.file, 'profPics')
+
         user.profilePic = image_url
         await user.save()
 
@@ -32,6 +38,7 @@ const deleteProfilePic = async (req, res) => {
         const user = await User.findOne({ username })
 
         const filepath = user.profilePic.split('/')
+        if ( filepath[filepath.length - 1] !== 'default')
 
         await deleteFileFromGoogleCS(filepath[filepath.length - 1], 'profPics')
